delete list

+++ app/src/main/java/com/nikolas/shoppinglist/db/MainViewModel.kt
@@ -44,8 +44,11 @@ class MainViewModel(database: MainDataBase) : ViewModel() {
         dao.deleteNote(id)
     }
 
-    fun deleteShopListName(id: Int) = viewModelScope.launch {
-        dao.deleteShopListName(id)
+    fun deleteShopList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+        if(deleteList) {
+            dao.deleteShopListName(id)
+        }
+        dao.deleteShopListItemsByListId(id)
     }
 
     class MainViewModelFactory(val database: MainDataBase) : ViewModelProvider.Factory {
